WIP: better changes

+++ src/interfaces/libpq-oauth/oauth-curl.c
@@ -1306,10 +1306,10 @@ register_socket(CURL *curl, curl_socket_t socket, int what, void *ctx,
 
 	return 0;
 #elif defined(HAVE_SYS_EVENT_H)
-	struct kevent ev[2] = {0};
-	struct kevent ev_out[2];
+	struct kevent ev[4] = {0};
+	struct kevent ev_out[4];
 	struct timespec timeout = {0};
-	int			nev;
+	int			nev = 0;
 	int			res;
 
 	/*
@@ -1320,24 +1320,52 @@ register_socket(CURL *curl, curl_socket_t socket, int what, void *ctx,
 	 * ENOENT is okay, but we have to track how many we get, so use
 	 * EV_RECEIPT.
 	 */
-	nev = 0;
 	EV_SET(&ev[nev], socket, EVFILT_READ, EV_DELETE | EV_RECEIPT, 0, 0, 0);
 	nev++;
 	EV_SET(&ev[nev], socket, EVFILT_WRITE, EV_DELETE | EV_RECEIPT, 0, 0, 0);
 	nev++;
 
-	Assert(nev == lengthof(ev_out)); /* otherwise we'll lose receipts! */
+	/* Add the new events. */
+	switch (what)
+	{
+		case CURL_POLL_IN:
+			EV_SET(&ev[nev], socket, EVFILT_READ, EV_ADD | EV_RECEIPT, 0, 0, 0);
+			nev++;
+			break;
 
-	res = kevent(actx->mux, ev, nev, ev_out, lengthof(ev_out), &timeout);
+		case CURL_POLL_OUT:
+			EV_SET(&ev[nev], socket, EVFILT_WRITE, EV_ADD | EV_RECEIPT, 0, 0, 0);
+			nev++;
+			break;
+
+		case CURL_POLL_INOUT:
+			EV_SET(&ev[nev], socket, EVFILT_READ, EV_ADD | EV_RECEIPT, 0, 0, 0);
+			nev++;
+			EV_SET(&ev[nev], socket, EVFILT_WRITE, EV_ADD | EV_RECEIPT, 0, 0, 0);
+			nev++;
+			break;
+
+		case CURL_POLL_REMOVE:
+			break;				/* nothing needs to be added */
+
+		default:
+			actx_error(actx, "unknown libcurl socket operation: %d", what);
+			return -1;
+	}
+
+	Assert(nev <= lengthof(ev));
+	Assert(nev <= lengthof(ev_out));
+
+	res = kevent(actx->mux, ev, nev, ev_out, nev, &timeout);
 	if (res < 0)
 	{
-		actx_error(actx, "could not delete from kqueue: %m");
+		actx_error(actx, "could not modify kqueue: %m");
 		return -1;
 	}
 
 	/*
 	 * We can't use the simple errno version of kevent, because we need to
-	 * skip over ENOENT while still allowing a second change to be processed.
+	 * skip over ENOENT while still allowing other changes to be processed.
 	 * So we need a longer-form error checking loop.
 	 */
 	for (int i = 0; i < res; ++i)
@@ -1352,57 +1380,25 @@ register_socket(CURL *curl, curl_socket_t socket, int what, void *ctx,
 		errno = ev_out[i].data;
 		if (!errno)
 		{
-			/* Successfully removed; update the event count. */
-			Assert(actx->nevents > 0);
-			actx->nevents--;
+			/* Successfully modified; update the event count. */
+			if (ev_out[i].flags & EV_ADD)
+				actx->nevents++;
+			else
+			{
+				Assert(ev_out[i].flags & EV_DELETE);
+				Assert(actx->nevents > 0);
+				actx->nevents--;
+			}
 		}
 		else if (errno != ENOENT)
 		{
-			actx_error(actx, "could not delete from kqueue: %m");
+			if (ev_out[i].flags & EV_DELETE)
+				actx_error(actx, "could not delete from kqueue: %m");
+			else
+				actx_error(actx, "could not add to kqueue: %m");
 			return -1;
 		}
 	}
-
-	/* If we're only removing registrations, we're done. */
-	if (what == CURL_POLL_REMOVE)
-		return 0;
-
-	/* Now add the new filters. This is more straightfoward than deletion. */
-	nev = 0;
-
-	switch (what)
-	{
-		case CURL_POLL_IN:
-			EV_SET(&ev[nev], socket, EVFILT_READ, EV_ADD, 0, 0, 0);
-			nev++;
-			break;
-
-		case CURL_POLL_OUT:
-			EV_SET(&ev[nev], socket, EVFILT_WRITE, EV_ADD, 0, 0, 0);
-			nev++;
-			break;
-
-		case CURL_POLL_INOUT:
-			EV_SET(&ev[nev], socket, EVFILT_READ, EV_ADD, 0, 0, 0);
-			nev++;
-			EV_SET(&ev[nev], socket, EVFILT_WRITE, EV_ADD, 0, 0, 0);
-			nev++;
-			break;
-
-		default:
-			actx_error(actx, "unknown libcurl socket operation: %d", what);
-			return -1;
-	}
-
-	res = kevent(actx->mux, ev, nev, NULL, 0, NULL);
-	if (res < 0)
-	{
-		actx_error(actx, "could not modify kqueue: %m");
-		return -1;
-	}
-
-	/* Update the event count, and we're done. */
-	actx->nevents += nev;
 
 	if (actx->debugging)
 		fprintf(stderr, "%s fd %d%s\n",
